Moved the testUtils file into the test directory, finalized the curl wrapper, and updated the Readme
