+++ upsurge/backend/app.js
@@ -37,7 +37,8 @@ app.post('/exp', function(req, res) {
     res.status(200);
     res.json(table);
   } catch (err) {
-    console.log(err)
+    console.log(err);
+    console.log(err.message);
     res.status(400);
     let msg = "There was an error parsing your boolean expression.\n";
     msg += "Make sure that your expressions are well formed, and that each connective is wrapped in parantheses with exactly two expressions on both sides.\n";
@@ -56,6 +57,8 @@ app.post('/circuit', function(req, res) {
     res.status(200);
     res.json(evaluation);
   } catch (err) {
+    console.log(err);
+    console.log(err.message);
     res.status(400);
     let msg = "There was an error parsing your circuit.\n";
     msg += "Make sure that you follow the specified format, and check the testing files for examples.\n";
