+++ upsurge/backend/wrapper.py
@@ -4,17 +4,19 @@ import requests
 
 if len(sys.argv) != 4:
     print("Incorrent command line arguments.")
-    print("python3 wrapper.py <local|remote> <circuit|expression> <expression string|path to circuit JSON file>")
+    print("python3 wrapper.py <local|remote> <circuit|expression> <path to JSON file>")
     quit()
 
 url = ""
 data = {}
-headers = {"Content-Type": "application/json"}
+headers = {"Content-Type": "application/json",
+           "Accept": "*/*",
+           "Connection": "keep-alive"}
 
 if sys.argv[1] == "local":
-    url = "localhost:3001"
+    url = "http://localhost:3001"
 elif sys.argv[1] == "remote":
-    url = "192.168.50.217:3001"
+    url = "http://192.168.50.217:3001"
 else:
     print("You did not specify a local or remote server!")
     quit()
@@ -30,5 +32,5 @@ else:
 with open(sys.argv[3]) as json_file:
     data = json.load(json_file)
 
-r = requests.post(url=url, headers=headers, data=data)
-print(r)
+r = requests.post(url=url, data=json.dumps(data), headers=headers)
+print(json.dumps(json.loads(r.text), indent=4, sort_keys=False))
