--- wrapperTest.json
@@ -1,3 +1,0 @@
-{
-    "expression": "not x1"
-}
